RE-1270 Remove PIP from setup_jenkins_slave.yml
As far as I can tell, there's no reason for this to happen any more.

This came to light testing the new rpc-openstack MNAIO jobs -- when run
through jenkins (where we install pip w/ rpc-gating's constraints file)
ansible failed to run with the following error:

```
ContextualVersionConflict: (cryptography 1.2.3 (/usr/lib/python2.7/dist-packages), Requirement.parse('cryptography>=1.5'), set(['paramiko']))
```

When testing manually (without installing pip w/ constraints), a newer
version of setuptools is installed and doesn't cause ansible to throw
that error.  The version of cryptography is still problematic but
ansible seems to be able to handle it internally and works as expected.