RE-1238 Add support for cloning internal Git repos
Cloning an internal repository must be done on an internal slave. To
facilitate this `clone_with_pr_refs` has been reworked to support both
internal and external repositories.

To prevent internal repository URLs from being exposed in this
repository, they are defined as a credential ID and so a repository URL
in a standard job can now take the form `"internal:some_credential_id"`.