RE-1238 Add new hardening jobs
The original hardening jobs have been removed because they were
implemented with legacy processes. This commit adds new hardening jobs
using the standard post-merge job template.

To enable this to work the project replicated the image and scenario
configuration from the equivalent rpc-openstack project definition. The
action combines the equivalent rpc-openstack project action with the
rpc-openstack branch to allow the rpc-openstack branch to be passed to
the gating scripts. By replicating the rpc-openstack configuration, the
hardening gating scripts are able to run the rpc-openstack gating
scripts with the same environment variables set.