RE-1300 Remove strict mode RPC-O MNAIO jobs for the RC branches
Given that the MNAIO tooling currently deploys the latest packages
for the base OS, the strict mode test will fail on the RC branches
as soon as a newer apt package than the artifacted packages releases.
Therefore, until a mechanism is worked out to deploy the base OS
using artifacts, the strict mode tests are just going to deliver
failures for a well known reason. As such, they should be removed
for now.