RE-1169 Specify container base image
In most cases, it's irrelevant which base image we use for creating a
container, however an ops-fabric test requires an ubuntu:17.10
container and subsequent tests call python3.6, which is not available
in the standard ubuntu base container image we use for all tests.

This review introduces the parameter `SLAVE_CONTAINER_BASE_IMAGE`,
allowing projects to specify which Ubuntu-flavored container base
image is used. Additionally, we duplicate the Dockerfile to
Dockerfile.standard_job because when the ARG is in the default
Dockerfile all jobs that use it which do not explicitly pass in
BASE_IMAGE will fail. This means this specific job will be impossible
to gate unless we manually update the docker call in Build-Gating-Venv.

NOTE: We remove the systemd-detect-virt call in
rpc_jobs/unit/single_use_slave.yml since systemd does not appear to be
installed on ubuntu:17.10.  Instead, we check for the existence of the
/.dockerenv file to determine if we're inside a container. This may not
be the most reliable long-term test since Docker may remove this file,
however for now it seems to work.