RE-1210 Fix skipping aborted-build issue creation
If a post-merge build is aborted by a user, the failure should not be
recorded in an issue. This change adds a new function to test if the
build was manually aborted, by looking for a specific interrupt action
as described in Jenkins issue JENKINS-28822 [1].

The code replaced in this commit does not work, the exception does not
have a `getCauses` method which results in `user` always being set to
`"SYSTEM"`.

[1] https://issues.jenkins-ci.org/browse/JENKINS-28822