(ASC-101) Create local jenkins for rpc-gating jobs
This adds a docker configuration to deploy a locally available
Jenkins instance. The jobs for rpc-gating can then deployed to this
instance using jenkins-jobs.

This is a three step process managed by the user and is documented in
the `jenkins_sandbox/README.md` file.

  * Build the Jenkins docker image
  * Run a docker container using the built image
  * Exec a `jenkins-job` command to deploy the rpc-gating jobs to the
  container.

The job deploy can be repeated as necessary while working on jjb code in
order to verify that the resulting job reflects the developers
intentions.