RO-3856 Remove legacy RPC-O archive collection
With the changes implemented in RPC-O to remove the use
of container artifacts, the job executing the container
builds is failing as it cannot find the appropriate log
folders on the host after the script executes. The script
is currently set to just exit immediately, and therefore
has nothing to collect.

In keeping with the rest of the jobs moving all log
collection to a standard location, then the job picking
up any artifacts if they're there, we remove the legacy
method of collection in this path to get those jobs moving
again and will add the standard log collection method
into the appropriate scripts as necessary. For now there
is actually little value in collecting anything other
than the console log anyway.