RE-1328 Fix job artifact upload to use the RE hook
Currently the job artifact upload assumes the directory
where the artifacts live, rather than making use of the
implemented hook to find out where it is.

This patch make the uploadToSwift function generic,
requiring only a path to find the stuff to upload. The
archive_artifacts function is then changed to make it
take parameters for the paths for the results/artifacts,
but to fall back to a set of default paths if they are
not provided.

The standard jobs are then changed to provide the correct
paths to the function so that the hooks are properly
implemented.