RE-1313 Use CRON parameter for PM jobs
In order to allow Post Merge (PM) job frequencies to be
configured, we implement the use of the already standard
"CRON" parameter.

This allows all standard periodic jobs to be configured
using the normal crontab syntax.

The default weekday CRON value is overridden for the
standard PM jobs to maintain the current frequency for
all current jobs.