(ASC-143) Add mnaio system test job
This commit adds 'system' as a action to
`rpc-openstack-master-mnaio-postmerge` project and a filter to ensure
that it only creates a single jobs for
`PM_rpc-openstack-master-xenial_mnaio_no_artifacts-swift-system`.

This job is will be the proof of concept job for running system level
testing on rpc-openstack. Hooks for taking advantage of the `system`
action have not yet been implemented in the `rpc-openstack` repo.