RE-36 Adjustable boot timeout & new snapshot jobs
Booting from a saved image is significantly slower than from a base
image, and as such we need to make the timeout configurable.

The default remains 900 seconds, but we can now override this via
BOOT_TIMEOUT.

Additionally, we add some new rpc-designate periodic jobs which use
snapshot images. These jobs are purely for testing and we will look
at adding process and moving all rpc-designate's jobs over to these
images once they are running reliably.

Lastly, we update a comment in the rpc-designate-post-merge project
which was made mention of the pre-merge template and not the
post-merge template.