RE-1345 Remove properties call in global wraps
Jobs have been mysteriously loosing their parameter definitions, and
my suspicion is that this call is overwriting all the properties set
in JJB rather than appending the log compression property to the list.

This may also be causing jobs to loose their github project affiliation
which could be contributing to problems with triggering.