RE-1372 Provide full ansible role source details
Ansible Galaxy is failing to download the 'mongrelion.docker'
role. It would appear that the author has done his first tagged
release and that ansible-galaxy does not allow a role download
that uses a SHA for a version when there is a tagged release.

To ensure that we're able to continue using the role at the
previous version, and to also provide enough information for
ansible-galaxy to fall back when the API service fails, this
patch adds the git source information for each role used.