RE-1390 Add jre 8 ppa on trusty builds
It looks like the SSH Slaves Plugin was upgraded from 1.15 to 1.26,
which apparently now requires jre 1.8 which is not available on Trusty.

As it's unclear if this plugin can be safely rolled back, I've
added a jre 8 ppa when building on trusty as a temporary workaround.

If we have confirmation that the plugin can be rolled back (and that
takes place) then we can revert this commit.