RE-1384 Make dep_update similar to standard jobs
The dep_update job is a standard job in the sense that it is designed to
be used by projects using pre/run/post scripts however the job
definition differs in a number of ways from the other standard jobs.
This change is a first pass at making the dep_update job closer in feel
to the other standard jobs.

The changes made by this commit include:
- switch the job from using long running slaves to containers to allow
  for system modification.
- renaming the job files to be consistent with other standard jobs.
- renaming the jobs to have the prefix PM-Dep-Update or PR-Dep-Update to
  be consistent with naming of other standard jobs.
- modifying the parameters/variables to have names consistent with the
  other standard jobs.
- the job name check in lint_jjb.py is modified to allow the first
  component to be templated.
- move rpc-openstack project for dependency updates to rpc_openstack.yml
  so that all rpc-openstack jobs are defined in the same file.