RE-1190 Use correct repo/ref for release PR tests
The current method of PR testing for releases uses
the remote fork for the test, instead of using the
target repo and the PR ref. This results in the test
failing in all kinds of exciting ways when the remote
fork does not contain the correct refs for the tooling
to be able do what it's supposed to do.

This patch ensures that the parameters passed to the
tooling are more appropriate and will ensure that all
the correct refs are there.