RE-1384 Add component dependency update support
This change adds support for testing component dependency updates. By
default this feature is not enabled but can be specified in the project
definition with `component_dependencies_update: "true"`. Currently this
feature is still being tested and should not be used outside of that
test work.

This change also allows for the use of update_dependencies/run updates
to be disabled with `third_party_dependencies_update: "false"`.