RE-1415 nodepool elements: Explicitly use jre 1.8
Jenkins requires the JRE 1.8 in order to connect to
the slaves. In this patch we ensure that the PPA is
setup if the image being built is on Ubuntu Trusty
(it has no native/backport package) and we also
ensure that the package installed afterwards is
explicitly set the the right one instead of assuming
the default package gives us JRE 1.8 for xenial.