RE-1422 Fix cleanup for long running nodes
Periodic cleanup is supposed to run a series of cleanups on all
long running slaves, one of which is to remove old docker containers.
However jobs started to fail as docker ran out of thin pool space.
On investigation it became clear that old docker containers were not
being removed. Executing the cleanup script manually successfully
removed old containers, and the cleanup job appeared to show
the script running correctly. The bug was that the cleanup script
was running N times on the same node, where N is the number of long
running slaves, rather than once against each node.

This is because of a variable scoping bug. The cleanup job loops over
the long running nodes in order to build a dictionary of tasks,
these tasks are then executed in parallel. However the nodeName variable
was globally scoped, so when the parallel tasks were executed, they
always used the last loop value. The remedy is to scope the nodeName
variable to within the loop its defined in, then each parallel
branch/closure will have a unique version of that variable.