RE-1449 Lint XML before running junit step
Currently, if bad XML ends up in the result dir and jenkins parses it
with the junit step, then jenkins will mark the build as UNSTABLE.

To avoid builds failing because of this situation, we first run xmllint
against the XML files and then only run the junit step if no errors
were found.

We also install libxml2-utils (which provides xmllint) in Docker
containers, pubcloud single use slaves, and nodepool instances.