RE-1317 Add release trigger
This commit adds a new `Component-Release-Trigger` job that is triggered
via comment to PRs against the rpc-metadata repo. When triggered, the
job will determine if a component is releasing and if so it will fire
off (in parallel) the component's `RELEASE_` jobs.

The trigger this job, you would simply add a comment with
`trigger-component-release` in it.

At present, the PR will need to be merged by the reviewer if/when the
`Component-Release-Trigger` job passes.  A subsequent PR will be
created to have the PR automatically merge if the job passes
successfully.