RE-1317 Add support for automated releases
This change extends Component-Release-Trigger to support automatically
releasing the requested version of the component and merging the pull
request.

Two new jobs are added to support the new functionality,
Component-Release runs RE-Release with the appropriate arguments. This
is do as part of Component-Release-Trigger if all the test jobs are
successful. If the release job is successful, Merge-Pull-Request is
triggered to merge the pull request.

Currently running this job will not result in the release being
performed however any pull requests will be merged.