RE-1316 RPC Gate Triggers
This commit adds a new job template for triggering component gate
checks.  Similar to the release gate, this is triggered manually via
the addition of a comment (`trigger-component-gate`) to the PR in
question. This should be added after all the initial checks have run and
the PR has been reviewed.  Upon successful completion of the gate jobs,
the PR will automatically merge.

We also modify the GATE_ template to add necessary parameters for
passing through PR details. We also remove the jira_project_key
reference in the stdJob() call to reflect what is currently done in
the pre-merge jobs. Lastly, we set a default for `branch` in the
template incase one doesn't get passed through by the project.