RE-1408 Sort /etc/hosts entries to improve idempotence
The zookeeper_hosts var contains a list which is different
on every playbook execution due to the way that the dynamic
inventory returns the data from OpenStack. This results in
/etc/hosts changing content on every playbook execution,
which causes all the services to be restarted every time.

By forcing the list to be sorted, we get a consistent result
back and the /etc/hosts file doesn't always change any more.
This results in an idempotent playbook execution.