RE-1317 Enable RE-Release in release pipeline
This commit alters the command supplied to RE-Release so that the
release job will perform the expected actions instead of simply printing
the command that would have been run.

`ORG` is updated to remove the assumption of rcbops and use the org
defined by the repo_url.

The use of optional is removed from the release notes run script,
components will be required to support this even if they simply have a
script to generate an empty file.