Don't skip entire junit step on bad XML
We now loop over XML files, running xmllint against each if xmllint is
installed. If a problematic file is found, we move it aside and run
the junit pipeline unconditionally against the remaining XML files.

If xmllint isn't installed, we just proceed as normal hoping that all
XML files are valid.