RE-1463 Rename g1-8 nodepool nodes and add p2-15 nodes
The functional PR tests all execute on nodes of two
flavor profiles: performance2-15, and general1-8.

This patch adds a distinct nodepool label for each
flavor profile so that they can be distinguished
from each other.

This patch also replaces the 'ubuntu-xenial' label
usage in jobs with 'ubuntu-xenial-g1-8', and the
'ubuntu-trusty' label usage in jobs with
'ubuntu-trusty-g1-8'.

The comments in the 'labels' are removed as the values
are no longer just for testing purposes, and the
'min-ready' argument is removed from the 'pool-labels'
as the argument does not apply there.