Revert "RE-1463 Revert g1-8 to working label and p2-15 to old config"
This reverts commit 935af07c6877001f480b820538be89eee1d94f0d.

Now that the issue of nodepool rejecting jobs for labels which
end with a number [1] is resolved, we can revert this change.

This extends the use of nodepool to more jobs and a broader set
of flavors, allowing us to monitor these for stability before
moving on to the next stage of job migrations.

[1] https://rpc-openstack.atlassian.net/browse/RE-1502