RE-1463 Use http for mirror.rackspace.com
Using https for the mirror is pointless, and prevents
caching from having any useful effect. A recent RPC-O
test failed due to a TLS error communicating with the
mirror.

In order to reduce the chances of failure, and improve
caching performance in tests, we switch the image to
use http instead of https.