RE-1489 Update issue link status-context
This commit updates the status-context for the issue link job to be
consistent with other status-contexts (underscore versus dash).

NOTE: Both rpc-openstack and rpc-gating have this job configured, but
neither have branches where this job is set to required. As such,
nothing should need to be done in GitHub once this merges.