RE-1464 Implement onmetal node config for nodepool
In order to manage the amount of onmetal nodes used for tests to ensure
that our quota is not exceeded (and tests fail due to that), we
implement the onmetal nodes in the nodepool configuration so that jobs
can use this method of obtaining the node instead of using the
single-use-slave method.

As we only have quota to do this in IAD, the configuration structure had
to be changed up a little to allow the new pool to be in IAD, but not
everywhere else.

Also, the build timeout had to be increased to cater for the onmetal
nodes.