RE-1464 Switch to using onmetal-io2
The fleet team has recommended that we move over to using onmetal-io2
instead of using onmetal-io2 as the io1 flavor will be phased out and
io2 is more readily available. They have also added quota for us in DFW
so that if IAD has issues, we have DFW to help keep test execution
going.

This patch implements the changes to the nodepool configuration and the
jobs to achieve this.