RE-1491 Update use of rpc_component
rpc_component has been moved from mattt416/rpc-metadata to
rcbops/rpc_component. This change updates the use of rpc_component to
pin it to the latest available version in rcbops/rpc_component. A
separate constraints file is added because rpc_component is Python 3
only and not compatible with the existing gating venv.

In addition, a few breaking changes have been addressed so that the new
version works where used. These changes centre around more consistent
use of the terms release and version.