RE-1466 Add nodepool config for RPC-O artifacted base images
In order to implement the ability for nodepool to build a base
image which is usable for RPC-O newton strict artifact builds,
we add the appropriate diskimage-builder elements and config
to build an image using the apt artifact repository, and the
appropriate nodepool config to ensure that we are able to
request nodes using this image.

We deliberately do not want to convert any jobs over to using
this image yet as we need to confirm that the nodes provided
with this image are usable and do not break the tests.