RE-1498 Browsable Artefacts
Currently all artefacts are compressed into a single archive and
uploaded to cloud files. This is convenient if you want to download all
the artefacts. However it turns out that most of the time it would be
more useful to be able to directly link to individual artefacts.

This commit changes upload artefacts so that each job has a unique
container and files can be browsed and Downloaded individually. Cloud
files has no index generation feature so a searchable index page is also
generated.

This commit also introduces batched archives. Artifacts are split
into batches of 1000 files for uploading for cloud files. These
archives also make it possile for a user to download a few files
in order to retrieve all the artifacts. A bash snippet is shown
on the index page to enable a user to easily grab all the artifacts.