RE-1487 Add component registration support
This change extends the standard release job to allow new components to
be added to the releases repo.

In addition, when a new component is registered this change creates a
pull request to rpc-gating to get the gate trigger job set up, the
trigger job is used to gate a component's pull requests.