RE-1545 Fix enable static website
The ARA report links to foo/ and expects foo/index.html to be returned.
That wasn't working, despite the rackspace docs saying:

  The page you set for X-Container-Meta-Web-Index becomes the index page
  for every subdirectory in your website.

Testing showed that this does work, but that upload_to_swift wasn't
enabling the static website correctly. The problem was that the
container name should have been appended to the object store url
used for enabling the static website.

This commit fixes this above problem and also enables file listing
so that a user can browse files/folders directly as well as using
the generated index.html.