RE-1486 Clean-up pull request branch on merge
This change updates the merge_pr function to attempt to delete the pull
request branch. If rpc-jenkins-svc has the correct permissions and the
branch isn't protected, it will delete the branch. This functionality is
used by the Merge-Pull-Request job.

The repo context object is modified to add the github session object to
allow interaction with the pull request repo because it may not be the
same as the base branch repo.