RE-1297 Remove `re` ref from status context
Currently the status context and trigger phrase doesn't match. Instead
of adding `re` to the trigger phrase, we just remove it from the status
context since we don't use this prefix anywhere else.

Additionally, we remove an unnecessary `?` in the trigger phrase. This
was left over when we removed the `cit` reference from the trigger
phrase in 36f717f5.

Once this merges, we'll need to update the repo settings to make this
test required (since it's going to appear as a new test).