RE-1466 Correct rpco-14.2-xenial-base flavor name
The flavor '7' isn't recognised by the shade library,
but using the actual name of the flavor works. This
patch corrects the name so that this base image is
usable, then sets the minimum number of available
nodes to verify that the image nodes start successfully
and are usable for testing.