RE-1513 Improve gate test guarantees
Without this change, any Component-Gate-Trigger_ job tests a pull
request against the target branch using the merge branch supplied by
GitHub but does not account for any changes that merge while those tests
are running.

This change ensures that a pull request is tested with any other pull
requests that are also going through the gate. To achieve this a build
looks for any builds that were started before it and includes their pull
requests in its own tests. If any of those builds fails, the tests on
dependent builds are terminated with a link to the upstream build and
new ones are started.

This change should provide a stronger guarantee as to the state of the
repository however if tests are unstable or pull requests are approved
that are broken, the time to merge a working pull request may increase.

This functionality can be disabled by setting `test_with_antecedents:
false` in a project.

Component-Gate-Trigger_ jobs that use this functionality no longer
depend on Merge-Pull-Request. This was done to simplify the tracking of
pull requests and the success of the tests. To enable merging to still
work, Component-Gate-Trigger_ will now update the status context for
"CIT/gate" so that the GitHub required status does not block the build
from merging the pull request.

REException class added to help reduce the noise in the build console
log. The new exception type is used to allow an exception to be raised
to exit the build without the stack trace appearing in the log.