RE-1484 Add rpc-openstack release jobs
This commit adds a master, pike and newton MNAIO release job for
rpc-openstack. The MNAIO jobs are often referred to to determine the
health of a branch before release, so it makes most sense to run these
prior to releasing.

Additionally, we update a comment in rpc_maas.yml to better reflect our
intention of splitting out check and gate jobs after rpc-maas has been
successfully added to the pipeline. This comment was inadvertently
copied from rpc-ceph and since rpc-maas already has a lint and
functional jobs this should be a straight-forward exercise for us to
do.