RE-1590 Add backup of default apt sources to rpco-artifacts image
The tests which make use of the artifacted image expect a backup
of the default apt sources to be in place to allow switching from
artifacted to non-artifacted setups. Instead of the tests having
to assume or derive the correct default sources we include a copy
in the image instead.