RE-1619 Catch Artifact Archival Failures
Artifact archival is not part of the test being executed, its a
convenience provided by RE. As such failures in that process
should not cause a build to fail.

This commit adds try/catch around artifact archival, and creates
a Jira issue on failure, instead of letting the exception propagate
up and fail the build.