ASC-665 Externalize RE_JOB_BRANCH for asc staging
This commit converts the asc staging job from a jjb project into a jjb
job in order to create the RE_JOB_BRANCH as an available parameter to be
set when a build is started. The RE_JOB_BRANCH is used internally by
the gating scripts to determine the branch of
`rpc-openstack-system-tests` to be checked out. Making this available as
a settable variable will allow topic branches of the testing repo to be
specified when triggering the job.