RE-1503 Add rpc-gating merge trigger job
In order to properly co-ordinate the execution of merge
trigger jobs when commits are merged into rpc-gating,
we create a single merge trigger job which co-ordinates
the others appropriately.

Instead of using a regex match, which may result in unexpected
jobs executing in the future, we opt for a tightly controlled
set of lists instead.

We also ensure that the job is serialised to ensure that all
jobs are executed and applied in the correct sequence when a
patch is merged.