RE-1503 Build gating venv once for unit tests
When a PR is submitted to rpc-gating, it currently starts
a bunch of unit test jobs, which in turn request a venv
build. Once the venv build is done, then they wait for the
repo sync, then execute. Instead of the indirect jobs request
in this patch we force the venv build to happen through one
blocking request. Once it's done it moves on to the parallel
unit test builds.

This helps reduce the load on jenkins and gets the job done
faster.