Define RE template for new components
This change adds the structure required by new components, this should
simplify the process of integrating into the RE process. When a new
component is added, a pull request is automatically proposed on the
component master branch which adds the expected files for integrating
with the RE framework. These can then be modified or deleted depending
on the project's specific requirements.

JIRA: RE-1529