Skip release tests that have already succeeded
If a release fails, this change ensures any RELEASE_ jobs, that need to
be run for the release, do not need to be re-run when the release is
retried from the same pull request if they succeeded and the SHA has not
changed. An example of where this helps is if the release process fails
when cloning rpc-gating from GitHub.

JIRA: RE-1684