ASC-662 Remove most leapfrog system jobs
This commit removes all but the following leapfrog system jobs.

* PM_rpc-upgrades-master-trusty_mnaio-swift-liberty_to_r14.current_leap_system
* PM_rpc-upgrades-master-trusty_mnaio-swift-liberty_to_newton_leap_system

At the time of this commit, all of the leapfrog system jobs were
consistently failing. The jobs would either fail during deployment or
the system tests would fail. In order to reduce the noise and not waste
infrastructure resources all but the above jobs are being deleted.

The above jobs should be enough for us to be able to resolve the root
cause of the failures. Once they are passing the jobs deleted in this PR
will be added back in a more controlled manner.

JIRA: ASC-662