Add retries and conditionals to cloudfiles upload
In order to improve the chances of the job artifact uploads
succeeding, all operations now have retries added, and more
specific conditions for success.

The rax_pub_cloud fact setting, related conditionals and tasks
for the 'not rax_pub_cloud' case are removed as that is an
unused and untested code path.

The 'Get Rackspace CloudFiles CDN URL' task is removed. The
previous task returns the data required, so this removes an
extra API call and point of failure.

The URI tasks have tightened response codes, with comments
for what each code means. This prevents failures leaking
through and makes it easier to understand what they are.

The URI tasks have 'no_log: True' removed, as they do not
expose any sensitive information unless the playbook is
run with '-vvv' which is not the case here.

As a matter of style and easier debugging, the following is
also done:

1. The vars are moved to the top of the playbook to make them
   more obvious when looking at the playbook.
2. All tasks are named according to what they do.
3. Comments are added/improved to make it clearer what is going
   on.
4. The service catalog data extraction is re-ordered to group
   the two object-store tasks together, and the two object-cdn
   tasks together.
5. All jinja vars are spaced. eg: "{{ object_cdn_url }}" instead
   of "{{object_cdn_url}}".
6. The 'Upload Artifacts to Cloud Files' command is split into
   multiple lines to make it easier to read, and is re-arranged
   to be in the same sequence as the other swift commands later
   in the play.

JIRA: RE-1618