RE-783 Change the label for build failure issues
The post merge test standard job (Created in RE-507) creates
JIRA issues on failure. These issues use the default label from
jirautils.py which is "jenkins-build-failure".  This task should allow a
label to be set in common.groovy (as "post-merge-test-failure" in
stdJob) and passes the label to jirautil.py.

JIRA: RE-783