Increase Cloud Files action retries/delays
The URI module tasks are still intermittently failing with
a 401 error (unauthorized), but sometimes pass a little later
on a retry. Before we go ahead and implement changes to the
os_object module, or try another region, we implement an
increase in delays and retries in the hope that an intermittently
failing API would not be doing so over the period of 5 mins
(10 retries with a 30 second delay between each).

JIRA: RE-1618