Verify ssh connectivity for jenkins slaves
Sometimes when an instance is created, it is not possible
to ssh to it. This is especially true of OnMetal nodes.

To try and prevent this from affecting tests before the
tests can be migrated to use nodepool, we implement a
try/rescue block so that if either the instance creation
or the ssh connection test fails, the instance is unusable
and a second attempt is done in the fallback region.

Also:

1. The variable 'primary_region_attempts' is removed, as it
   is unused.
2. The variable 'regions_shuf' is renamed to 'regions_shuff'
   so that it is more uniform with 'fallback_regions_shuff'.
3. The difference filter used in 'fallback_regions_shuff' is
   documented, and fixed to use 'regions_shuff' as it should
   be.
4. Extra debug output is added to help diagnose issues which
   may arise with this approach.
5. If there is no valid region list the tasks will fail early
   instead of trying to execute and resulting in an obscure
   error.

JIRA: RE-1517