Update RE jobs to set OS_CLOUD when creating instances
When Ansible uses the dynamic inventory to create instances, the
inventory currently fails due to the new Phobos cloud configuration
which is unreachable for the Jenkins slaves.

This patch updates the dynamic inventory to the latest available
version in 'stable-2.4' branch which provides the ability to set
a specific OS_CLOUD environment variable to tell the script to only
look at that cloud.

We also update all the jobs to make use of this environment variable.

JIRA: RE-1751