Separate the apt commands for OnMetal tests
To ensure that we're able to effectively handle transient
failures, we split the apt cache update and installs, and
we add retries to both. We also shift to using 'apt-get'
instead of 'apt' as the output does not have color coding
and is therefore easier to read in Ansible's output.

JIRA: RE-1752