Add debug outputs to help diagnose apt issues
Some of the OnMetal hosts are failing to install the
python-minimal and python-yaml packages when they are
initially built, but the reason why they're failing
is not clear.

This patch adds some diagnostic configuration to the
host and some debug outputs to the tasks to help figure
out what's going on.

JIRA: RE-1758