RE-1655 Attempt to read broken xml
Set the recover flag on the lxml parser so that it attempts to continue
when it comes across an invalid character.

The Jenkins build failure analysis plugin inserts raw excerpts of job
logs into the build.xml, which may cause it to contain invalid
characters. Before this commit, all those builds would be skipped
by build summary as build.xml parsing fails.