Reconfigure apt sources for OnMetal hosts
The OnMetal hosts have their images prepared by Ubuntu
and therefore include the default use of Ubuntu mirror
for apt packages. These have turned out to be a bit
flaky so we have opted to reconfigure the hosts with
the Rackspace mirror instead, which we've found in
all our tests to be reliable.

This patch implements that change in configuration, but
also ensures that this is only done on a host which does
not already have python installed. This prevents any
unnecessary execution of the reconfiguration tasks.

JIRA: RE-1758