RE-1767 Ensure inventory stash happens
If an error happens during the playbook that allocates instances,
then groovy throws an exception and the stash task doesn't happen.
If the stash doesn't happen then cleanup fails, as cleanup relies
on retrieving an inventory file from a stash.

So if the playbook succesfully allocates an instance, then fails,
the instance will be leaked because cleaup will fail.

This change adds try/finally to ensure that stash is always
attempted regardless of playbook success.